% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{BallMapper}
\alias{BallMapper}
\title{Create vertices and edges (with additional properties) of a Ball Mapper graph of the input data.}
\usage{
BallMapper(points, values, epsilon)
}
\arguments{
\item{points, }{a collection of input points in a form of a data frame. Those are typically points in Euclidean space.}

\item{values, }{a collection of function values on points. This variable will be used to color the Ball Mapper graph. If it is not available, please set it up to a constant array.}

\item{epsilon, }{the value of radius of balls used in the Ball Mapper construction.}
}
\value{
The function return a long list of parameters that are explained below:\
vertices, which consist of two binded list: first one contains an increasing sequence of numbers starting from 1 to the number of vertices. Each of them correspond to a landmark point. The second one contains the number of points colvered by a ball of radius epsilon centered by following landmark points.
edges, a collection of not directed edges composed of the first and the second vertex. Ordering of vertices do not have meaning.
edges_strength, For every edge [a,b] we define its strength as the nuber of points that are both covered by landmarks a and b. This array contains the strengt of every edge in the Ball Mapper graph.
points_covered_by_landmarks, is a list of vectors. I-th vector contains the positions of points covered by i-th landmark.
landmarks, contains a list of positions of landmark points
coloring, is a vector having as many positions as number of lanrmarks. It contains the averaged values of points covered by each landmark.
}
\description{
Create vertices and edges (with additional properties) of a Ball Mapper graph of the input data.
}
