% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{coloredDynamicNetwork}
\alias{coloredDynamicNetwork}
\title{This procedure produces a dynamic graph with colors. It allows zoom-in operation and display information on vertices when clicked on them.}
\usage{
coloredDynamicNetwork(outputOfBallMapper, showLegend = FALSE)
}
\arguments{
\item{showLegend, }{if set to TRUE will display the legend indicating the coloring of the values of vertices.}

\item{outputFromBallMapper, }{an output from BallMapper function}
}
\description{
This procedure produces a dynamic graph with colors. It allows zoom-in operation and display information on vertices when clicked on them.
}
