% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{ColorIgraphPlot}
\alias{ColorIgraphPlot}
\title{Produce a static color visualization of the Ball Mapper graph. It is based on the output from BallMapper function.}
\usage{
ColorIgraphPlot(outputFromBallMapper, showVertexLabels = TRUE,
  showLegend = FALSE, maximal_ball_scale = 20,
  maximal_color_scale = 10)
}
\arguments{
\item{outputFromBallMapper, }{an output from BallMapper function}

\item{showVertexLabels, }{a boolean value determining if vertex labels are to be shown (TRUE by default).}

\item{showLegend, }{a boolean value determining if the legend is to be shown (FALSE by default).}

\item{maximal_ball_scale, }{provide a maximal value of a radius of a ball used in visualization (20 by default).}

\item{maximal_color_scale, }{Provide a maximal value (starting from 0) of a color of a ball (10 by default).}
}
\description{
Produce a static color visualization of the Ball Mapper graph. It is based on the output from BallMapper function.
}
