% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{GrayscaleIgraphPlot}
\alias{GrayscaleIgraphPlot}
\title{Produce a static grayscale visualization of the Ball Mapper graph. It is based on the output from BallMapper function.}
\usage{
GrayscaleIgraphPlot(outputFromBallMapper, showVertexLabels = TRUE,
  maximal_ball_scale = 20)
}
\arguments{
\item{outputFromBallMapper, }{an output from BallMapper function}

\item{showVertexLabels, }{a boolean value determining if vertex labels are to be shown (TRUE by default).}

\item{maximal_ball_scale, }{provide a maximal value of a radius of a ball used in visualization (20 by default).}
}
\description{
Produce a static grayscale visualization of the Ball Mapper graph. It is based on the output from BallMapper function.
}
