% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{simpleDynamicNetwork}
\alias{simpleDynamicNetwork}
\title{This is a simple example of dynamic visualization using networkD3 library.
This version do not implement coloring of vertices, just give a general overview of the edges.}
\usage{
simpleDynamicNetwork(l)
}
\arguments{
\item{outputFromBallMapper, }{an output from BallMapper function.}
}
\description{
This is a simple example of dynamic visualization using networkD3 library.
This version do not implement coloring of vertices, just give a general overview of the edges.
}
